move header optics to Data.RFC5322
Headers that are defined in RFC 5322 should be implemented in and
exported from module Data.RFC5322.  Make it so.
